+++ CHANGELOG.md
@@ -13,7 +13,8 @@ phpSysInfo 3.2.9
 
  - [NEW] SHOW_DEVICES_INFOS and SHOW_DEVICES_SERIAL options
  - [NEW] Thinkpad - Lenovo Thinkpad sensors monitoring for Linux
- - [NEW] StableBit plugin
+ - [NEW] StableBit plugin - StableBit Scanner info
+ - [NEW] HyperV plugin - Hyper-V info
 
 phpSysInfo 3.2.8
 ----------------